requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai==1.3.5
python-docx==1.1.0
PyMuPDF==1.23.8
python-dotenv==1.0.0
mangum==0.17.0

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-mock==3.12.0
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
pre-commit==3.6.0
psutil==5.9.6